+++ django_gotify/__init__.py
@@ -1,0 +1,2 @@
+from .email import GotifyEmailBackend
+from .log import GotifyLogHandler