+++ django_gotify/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class DjangoGotifyConfig(AppConfig):
+    name = "django_gotify"
+    verbose_name = "Django Gotify"